reused FactionFlag in VictoryPoints
